GraphqlTemplateFastApi/requirements.txt
fastapi==0.79.1
uvicorn[standard]==0.18.2
strawberry-graphql[fastapi]==0.126.1
gunicorn==20.1.0
reactivex==4.0.4
py-healthcheck==1.10.1
alembic==1.8.1
requests==2.28.1
pydantic==1.9.2
python-graphql-client==0.4.3
paho-mqtt==1.6.1
pika==1.3.0
PikaBus==1.3.6
